systemd: removed enlightenment-user.path
enlightenment-user.service wil be activated when starting basic.target.wants
and check itself whether /run/wayland-0 socket file exists using shell scripts
in ExecStart.

If the path-based activation is used, then enlightenment-user.service is not
registered as a job of the systemd. In this case, the systemd could ignore
dependencies on the enlightenment-user.service.

Change-Id: Icc244608dfd7653213f931a506e88059faf7d235